Convert from array to obj-oriented task creation.
Previously, code depended on the used inputs being in the forced correct
order, and the card would populate the info in that order. Now the user
info is stored in a obj where each prop is connected to it's individual
card line. Allows for reordering of the card's info lines.

Next is to add date filtering (similar to project filtering).