Add "add project" btn to project list.
This function works similar to the edit button on the todo card lines,
at least far as hiding the line/button and bringing up a confirm/cancel
option. So I moved the edit button's logic to a seperate function inside
of render.js, and allowed it to accept a callback function for the
confirm buttons logic. The call backs are written out their respective
funcitons in eventListeners.js.

Still need the projects drop down menu in the default todo card.