Add edit project button to project list in sidebar.
The side bar project list is created by looping through the current
tasks and extracting the project names. So this function just updates the
project name of any task in the user's selection to the new user project
project name and re-renders as normal. Allows user to combine tasks into
one project.