Fix add project bug (from edit project function).
Cannot wipe the project list on render all. This deletes the project
that the used just added because the project list is generated from the
task list (therre are no task in the new project yet so its not
included in the project list). The fix instead deletes the edited project
out of the current project list, allowing empty "projects" to exist.