Fix bug: editing empty projects deletes projects
Added line to the "edit project confirm button" that manually adds the
updated project name to the project list, versus only deleting the old
and depending on the tasks to create the project list (why "empty"
project were being deleted: the edit deleted the old name and there were
no tasks with the new project name to add the project to the project
list). There is a blend of everything depending on the tasks list and
adding projects, making projects more real and less of an "illusion".