Add delete button to project list.
Borrowed code from edit button. Modified to not include a new input but
rather the name crossed out. On confirmation, all tasks with that project
name are deleted. The project list is then created from the task list
(deleted project now has no tasks and is, thus, no longer in the list).
Will probably find a bug, per the norm.