Last commit before restarting.
Being that I was interrupted in this project for a few months and seeing
some lacks in using modules, I feel it would be good to redo the lessons
leading up to this project and then do the project again, focusing on
the bare-bones of what is being asked. This allows me to refresh the
concepts I learned months ago, correct my mistakes, and move forward.